Fixed implicit casts
Problem detected by Clang's better support for the -Wconversion flag.

git-svn-id: http://130.66.124.6/svn/maroff/Product/Cpp/Tools/random_data_generator@1711 300066c3-8af3-463d-8cc0-c0233cc64ac1

+++ random_data_generator/src/DataGenerator.cpp
@@ -86,7 +86,7 @@ char DataGenerator::random_char() const
             "0123456789"
             "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
             "abcdefghijklmnopqrstuvwxyz";
-    const size_t random_index = get_random_number() % (sizeof(alphanum)/sizeof(char) - 1);
+    const size_t random_index = (size_t)get_random_number() % (sizeof(alphanum)/sizeof(char) - 1);
     return alphanum[random_index];
 }
